+++ scripts/update-schema-toml.sh
@@ -1,8 +1,0 @@
-#!/usr/bin/env bash
-
-set -exuo pipefail
-cd "$(dirname "$0")/.."
-
-mkdir -p assets/schemas/toml
-cargo run -- generate server-config-schema assets/schemas/toml/server.json
-cargo run -- generate deployment-schema assets/schemas/toml/deployment.json